requirements.txt
faust-streaming
boto3